// server/src/types/fields-of-mistria/game-stats.d.ts
export interface FieldsOfMistriaSetCompletion {
  day: number;

  wing: string;

  set: string;
}

export interface FieldsOfMistriaItemsSoldLog {
  income: number;

  count: number;

  item: string;
}
// ...
export interface FieldsOfMistriaIncome {
  amount: number;

  day: number;

  type: string;
}
// ...
export interface FieldsOfMistriaRenownLevelUp {
  level: number;

  day: number;
}